optimisation+correction

+++ src/NinjaTooken/UserBundle/Controller/DefaultController.php
@@ -463,9 +463,12 @@ class DefaultController extends Controller
                     $extension = strtolower($file->guessExtension());
                     if(in_array($extension, array('jpeg','jpg','png','gif'))){
                         $user->file = $file;
-                        $cachedImage = dirname(__FILE__).'/../../../../web/cache/avatar/'.$user->getWebAvatar();
-                        if(file_exists($cachedImage)){
-                            unlink($cachedImage);
+                        $userWebAvatar = $user->getWebAvatar();
+                        if(isset($userWebAvatar) && !empty($userWebAvatar)){
+                            $cachedImage = dirname(__FILE__).'/../../../../web/cache/avatar/'.$userWebAvatar;
+                            if(file_exists($cachedImage)){
+                                unlink($cachedImage);
+                            }
                         }
                         $user->setAvatar('');
                     }
